[refactor] 배틀에 사용되는 큐와 맵을 분리
- inMemory component로 큐와 맵을 관리하도록 함
- match와 skill 서비스 로직을 분리하기 위함
- 배틀 로직 변경 전 분리가 필요해서 함
