+++ BE/src/main/java/cuketmon/battle/controller/BattleWebSocketController.java
@@ -3,6 +3,7 @@ package cuketmon.battle.controller;
 import cuketmon.battle.dto.SkillRequest;
 import cuketmon.battle.dto.TrainerRequest;
 import cuketmon.battle.service.BattleMatchService;
+import cuketmon.battle.service.BattleSkillService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
@@ -12,10 +13,12 @@ import org.springframework.stereotype.Controller;
 public class BattleWebSocketController {
 
     private final BattleMatchService battleMatchService;
+    private final BattleSkillService battleSkillService;
 
     @Autowired
-    public BattleWebSocketController(BattleMatchService battleMatchService) {
+    public BattleWebSocketController(BattleMatchService battleMatchService, BattleSkillService battleSkillService) {
         this.battleMatchService = battleMatchService;
+        this.battleSkillService = battleSkillService;
     }
 
     // 배틀 찾기 (대기열 등록)
@@ -34,7 +37,7 @@ public class BattleWebSocketController {
     // 기술 사용
     @MessageMapping("/skill/{battleId}")
     public void useSkill(@DestinationVariable Integer battleId, SkillRequest skill) {
-        battleMatchService.useSkill(battleId, skill);
+        battleSkillService.useSkill(battleId, skill);
     }
 
 }
